+++ src/utils/hooks.ts
@@ -1,0 +1,23 @@
+import { useEffect, useMemo, useState } from "react";
+import { TEMPLATES } from "./text";
+
+export const useTemplateRotation = (msg: string | null) => {
+  const [templateIdx, setTemplateIdx] = useState<number>(
+    Math.floor(Math.random() * TEMPLATES.length),
+  );
+  const msgShown = useMemo(() => {
+    return msg || TEMPLATES[templateIdx];
+  }, [msg, templateIdx]);
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setTemplateIdx(Math.floor(Math.random() * TEMPLATES.length));
+    }, 5000);
+
+    return () => {
+      clearInterval(interval);
+    };
+  }, []);
+
+  return msgShown;
+};
